MicroService_Project/app/requirements.txt
flask
redis
mysql-connector